// packages/api/discovery/application/models/SanityTypes.ts
export type CompanyLocation = {
  _type: 'company.location';
  location?: Geopoint;
  storeName?: string;
  addressLine1?: string;
  addressLine2?: string;
  townCity?: string;
  region?: string;
  postcode?: string;
  country?: 'UK' | 'AU';
  telephone?: string;
};

export type Geopoint = {
  _type: 'geopoint';
  lat?: number;
  lng?: number;
  alt?: number;
};

export type TagCategory = {
  _type: 'tag.category';
  tagCategoryName?: string;
  tags?: Array<Tag>;
};

export type Attribute = {
  _type: 'attribute';
  name?: string;
  description?: string;
};

export type Restriction = {
  _id: string;
  _type: 'restriction';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  description?: string;
};

export type AgeRestriction = {
  _id: string;
  _type: 'age.restriction';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  description?: string;
};
// ...
export type TrustService = {
  _id: string;
  _type: 'trust.service';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  description?: string;
  code?: string;
};
// ...
export type Company = {
  _id: string;
  _type: 'company';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  brandCompanyDetails?: Array<{
    brand?: Brand;
    companyName?: string;
    displayName?: string;
    companyShortDescription?: Array<{
      children?: Array<{
        marks?: Array<string>;
        text?: string;
        _type: 'span';
        _key: string;
      }>;
      style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote';
      listItem?: 'bullet' | 'number';
      markDefs?: Array<{
        href?: string;
        _type: 'link';
        _key: string;
      }>;
      level?: number;
      _type: 'block';
      _key: string;
    }>;
    companyId?: number;
    legalPrivacyLink?: string;
    companyLogo?: {
      default?: {
        asset?: SanityImageAsset;
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        _type: 'image';
      };
      light?: {
        asset?: SanityImageAsset;
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        _type: 'image';
      };
      dark?: {
        asset?: SanityImageAsset;
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        _type: 'image';
      };
    };
    ageRestrictions?: Array<AgeRestriction>;
    _key: string;
  }>;
  legalEntityName?: string;
  alsoKnownAs?: Array<string>;
  restrictions?: Array<TrustService>;
  local?: boolean;
  companyLocations?: Array<
    {
      _key: string;
    } & CompanyLocation
  >;
  categorySelection?: Array<{
    category1?: Category;
    category2?: Category;
    category3?: Category;
    categoryItem?: CategoryItem;
    _key: string;
  }>;
  eagleEyeMerchantId?: string;
  suspended?: boolean;
};

export type CategoryItem = {
  _id: string;
  _type: 'category.item';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  id?: number;
};

export type Category = {
  _id: string;
  _type: 'category';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  id?: number;
  level?: number;
  parentCategoryIds?: Array<number>;
  categoryItems?: Array<CategoryItem>;
};
// ...
export type Brand = {
  _id: string;
  _type: 'brand';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  link?: string;
  code?: string;
  name?: string;
  description?: string;
  logo?: {
    default?: {
      asset?: SanityImageAsset;
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: 'image';
    };
    light?: {
      asset?: SanityImageAsset;
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: 'image';
    };
    dark?: {
      asset?: SanityImageAsset;
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: 'image';
    };
  };
};

export type SanityImageCrop = {
  _type: 'sanity.imageCrop';
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot';
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageAsset = {
  _id: string;
  _type: 'sanity.imageAsset';
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  originalFilename?: string;
  label?: string;
  title?: string;
  description?: string;
  altText?: string;
  sha1hash?: string;
  extension?: string;
  mimeType?: string;
  size?: number;
  assetId?: string;
  uploadId?: string;
  path?: string;
  url?: string;
  metadata?: SanityImageMetadata;
  source?: SanityAssetSourceData;
};

export type SanityAssetSourceData = {
  _type: 'sanity.assetSourceData';
  name?: string;
  id?: string;
  url?: string;
};

export type SanityImageMetadata = {
  _type: 'sanity.imageMetadata';
  location?: Geopoint;
  dimensions?: SanityImageDimensions;
  palette?: SanityImagePalette;
  lqip?: string;
  blurHash?: string;
  hasAlpha?: boolean;
  isOpaque?: boolean;
};